refactor(board) : getBoard.ts -> board.ts로 변경, server fetching으로 변경
